==> calculation/20_ordination.R <==
# Ordination
# Sean Kinard
# 2023-06-23
#------------------------------------------------------------------------------
# setup
#-----------------------------------------------------------------------------
source('exploration/toolkit.R')
library(BiodiversityR)

d_density <- read_csv("data/06_fill_data/fish_density_fill.csv")
#d_biomass <- read_csv("data/06_fill_data/fish_biomass.csv")
d_bio_categories <- read_csv("data/02_clean_data/fish_biological_scales_of_comparison.csv")
d_ste <- read_csv("exploration/output/environment_shortterm_with_ln.csv")

RDA_evars <- c("ln_q_med", "ln_q_rsd", "q_HFPP3",
               "conductivity", "nitrate", "phosphate", "d_oxygen",
               "gravel", "silt", "temperature", "ln_rain", "total_algae")

#------------------------------------------------------------------------------
# Pre-Processing Data
#------------------------------------------------------------------------------
d_both <- left_join(d_density, d_ste) 

# RDA Preparation: Species
d_spe <- d_both %>%
  unite('UID', site_code, collection_period, sep='_') %>%
  #left_join(select(d_bio_categories, lowest_taxon, total_occurances)) %>%
  #filter(total_occurances >= 10) %>%
  select(density, lowest_taxon, UID) %>%
  mutate(density = ceiling(density*1000)) %>% # transform to pseudo-abundance
  pivot_wider(names_from=lowest_taxon,
              values_from=density, 
              values_fill = 0) %>%
  column_to_rownames(var="UID")

# fill missing predictors with linear interpolation
d_ste_fill <- d_both %>%
  rowwise() %>%
  mutate(total_algae = sum(sqrt_green_algae^2, 
                           sqrt_diatoms^2, 
                           sqrt_phosphate^2)) %>%
  select(site_code, collection_period, any_of(RDA_evars)) %>%
  unique() %>%
  as_tsibble(key=site_code,
             index=collection_period) %>%  
  na_ma(weighting="linear") %>%
  ungroup() %>%
  as_tibble() %>%
  pivot_longer(cols=-any_of(c('site_code', 'collection_period')),
               names_to='xname',
               values_to = 'x') %>%
  group_by(site_code, collection_period, xname) %>%
  dplyr::summarize(x=mean(x ,na.rm=T)) %>%
  pivot_wider(names_from=xname, values_from = x) %>%
  ungroup()

# RDA preparation: Environment
d_predictors <- d_ste_fill %>%
  unite('UID', site_code, collection_period, sep='_') %>%
  column_to_rownames("UID")

d_both %>%
  unite('UID', site_code, collection_period, sep='_') %>%
  #left_join(select(d_bio_categories, lowest_taxon, total_occurances)) %>%
  #filter(total_occurances >= 10) %>%
  select(density, lowest_taxon, UID) %>%
  mutate(density = ceiling(density*1000)) %>% # transform to pseudo-abundance
  pivot_wider(names_from=lowest_taxon,
              values_from=density, 
              values_fill = 0) %>%
  pivot_longer(cols=-UID, names_to = 'xvar', values_to = 'x') %>%
  filter(is.na(x))

# ------------------------------------------------------------------------------
# RDA Calculations
#------------------------------------------------------------------------------
# function runs RDA and creates a list of data frames for plots and tables
my_rda <- function(d_env, d_spec) {
  
  # step 1: The ordiplot object is obtained from the result of via function rda. The ordination is done with a community data set that is transformed by the Hellinger method.
  # script generated by the BiodiversityR GUI from the constannualrained ordination menu
  fish_hell <- disttransform(d_spec, method='hellinger')
  Ord_model <- rda(fish_hell ~ . ,data=d_env, scaling="species")
  # This can be plotted via ordiplot
  plot2 <- ordiplot(Ord_model, choices=c(1,2))
  #------------------------------------------------------------------------------
  # Step 2 : To plot data via ggplot2
  # Extract information on the locations of sites (circles in the ordiplot)
  sites_long <- sites.long(plot2, env.data=d_env) %>%
    mutate(site_code = substr(rownames(d_env), 1,2),
           collection_period = substr(rownames(d_env), 4,13),
           collection_period = ymd(collection_period),
           cd = collection_period) %>%
    separate(cd, into=c('year', 'month', 'day'), sep='-')
  # Extract species info
  species_long <- species.long(plot2) %>%
    as_tibble() %>%
    mutate(labels = str_replace_all(labels, '\\..', '.'))
  # Information on the labeling of the axes is obtained with function axis.long. This information is extracted from the ordination model and not the ordiplot, hence it is important to select the same axes via argument 'choices'. The extracted information includes information on the percentage of explained variation. Be advised to cross-check with the summary of the redundancy analysis, where information on proportion explained is given.
  axis_long <- axis.long(Ord_model, choices=c(1, 2))
  #------------------------------------------------------------------------------
  # fit species vectors
  spp_fit <- envfit(plot2, env=fish_hell)
  df_spp_fit <- data.frame(r=spp_fit$vectors$r, p=spp_fit$vectors$pvals)
  spec_vector_long <- species.long(plot2, spec.data=df_spp_fit)
  RDA_sv <- spec_vector_long[spec_vector_long$p < 0.05, ] %>%
    as_tibble() %>% mutate(labels = str_replace_all(labels, '\\..', '.'))
  
  # fit environmental vectors
  vectors.envfit <- envfit(plot2, env= d_env)
  vectors.long2 <- filter(vectorfit.long(vectors.envfit), p < 0.05)
  RDA_ev <- vectors.long2[vectors.long2$p <= 0.05, ]
  #------------------------------------------------------------------------------
  # Table
  RDA_table <- species_long %>%
    dplyr::rename(vector = labels) %>%
    full_join(vectors.long2) %>% 
    filter(p < 0.05) %>%
    mutate(r = round(r, 3),
           p = round(p, 3),
           axis1 = round(axis1, 3),
           axis2 = round(axis2, 3)) %>%
    select(vector, r, p, axis1, axis2) %>%
    dplyr::rename('p-value'=p, 'Axis-1'=axis1, 'Axis-2'=axis2, Vector=vector)
  
  # ---------------------------------------------------------------------------
  # Format time variables
  sites_long <- sites_long %>%
    mutate(year=year(collection_period),
           month=month(collection_period),
           day=day(collection_period),
           nday=yday(collection_period),
           qtr = quarter(collection_period),
           year_week = yearweek(collection_period),
           year_month = yearmonth(collection_period)) %>%
    mutate(days_to_hh = 
             as.numeric(as_date(collection_period) - ymd('2017-08-27')) ) %>%
    mutate(weeks_to_hh = ceiling(days_to_hh/7),
           months_to_hh = round(days_to_hh/30, 0)) %>%
    mutate(site_week_label = paste(site_code, weeks_to_hh, sep="-"))
  
  # -----------------------------------------------------------------------
  # Calculate all-time centroid for each site
  mycent_alltime <-  sites_long %>%
    add_rain() %>%
    group_by(annualrain) %>%
    dplyr::summarize(axis1_mu = mean(axis1),
                     axis2_mu = mean(axis2),
                     axis1_upr = quantile(axis1, probs=.9),
                     axis1_lwr = quantile(axis1, probs=.1),
                     axis2_upr = quantile(axis2, probs=.9),
                     axis2_lwr = quantile(axis2, probs=.1))
  
  labs_alltime <- mycent_alltime %>%
    select(annualrain, axis1_mu, axis2_mu) %>%
    mutate(annualrain_label = paste(round(annualrain, 0), 'mm', sep=' ')) %>%
    unique()
  # -----------------------------------------------------------------------
  # centroid distance
  centroid_distance <- sites_long %>%
    add_rain() %>%
    as_tibble() %>%
    select(site_code, annualrain, collection_period, axis1, axis2) %>%
    left_join( mycent_alltime %>% select(annualrain, axis1_mu, axis2_mu) ) %>%
    mutate(dif_x1 = axis1-axis1_mu,
           dif_x2 = axis2-axis2_mu,
           centroid_dist = sqrt(dif_x1^2 + dif_x2^2)) %>%
    select(site_code, collection_period, centroid_dist)
  # -----------------------------------------------------------------------
  RDA_out <- list(axis_long = axis_long,
                  sites_long = sites_long,
                  species_long = species_long,
                  RDA_ev = RDA_ev,
                  RDA_sv = RDA_sv,
                  RDA_table = RDA_table,
                  mycent_alltime = mycent_alltime,
                  labs_alltime = labs_alltime,
                  centroid_distance = centroid_distance)
  
  return(RDA_out) 
}

#------------------------------------------------------------------------------
# RDA environment plot function
#------------------------------------------------------------------------------
RDA_env <- function(rda_out) {
  # -----------------------------------------------------------------------
  # Index items from rda_output
  axis_long <- rda_out[[1]] %>% as_tibble()
  sites_long <- rda_out[[2]] %>% as_tibble() %>% fix_site_order() %>% add_rain()
  species_long <- rda_out[[3]] %>% as_tibble()
  RDA_ev <- rda_out[[4]] %>% as_tibble()
  RDA_sv <- rda_out[[5]] %>% as_tibble()
  mycent_alltime <- rda_out[[7]] %>% as_tibble()
  labs_alltime <- rda_out[[8]] %>% as_tibble()
  
  # ---------------------------------------------------------------------------
  # environmental Plot
  ggplot() +
    #geom_rect(aes(ymin=axis2_lwr, ymax=axis2_upr,
    #              xmin=axis1_lwr, xmax=axis1_upr,
    #              fill=annualrain),
    #          data= mycent_alltime, 
    #          color=NA, alpha = .25) +
    #stat_ellipse(data = sites_long,
    #aes(x=axis1, y=axis2, fill = as.factor(annualrain)),
    #geom='polygon', alpha=.1, level = 0.9,show.legend = F) +
    #stat_ellipse(data = sites_long,
    #aes(x=axis1, y=axis2, color = as.factor(annualrain)),
    #alpha=.5, level = 0.9, show.legend = F) +
  geom_segment(data=RDA_ev,
               aes(x=0, y=0, xend=(3*r)^2*axis1, yend=(3*r)^2*axis2),
               colour="grey90", linewidth=.5, 
               arrow=arrow(length=unit(0.1, "inches"),
                           type='open')) +
    geom_label(aes(label=annualrain_label,
                   x=axis1_mu, y=axis2_mu, fill = as.factor(annualrain)),
               data= labs_alltime, show.legend=T, color ='black' ) +
    #geom_point(data=species_long,
    #aes(x=axis1, y=axis2), 
    #color = 'white', fill='black', shape = 24, size = 2) +
    
    geom_text_repel(
      data=RDA_ev,
      aes(x=(3*r)^2*axis1, y=(3*r)^2*axis2, label=vector),
      color="white", fill = 'grey15', box.padding = .9,
      max.overlaps = 15) +
    xlab(axis_long[1, "label"]) +
    ylab(axis_long[2, "label"]) +
    scale_x_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
    scale_y_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
    scale_fill_manual(values = my_colors) +
    scale_color_manual(values=my_colors) +
    # paletteer::scale_fill_paletteer_c("grDevices::Zissou 1", direction=-1) +
    # paletteer::scale_color_paletteer_c("grDevices::Zissou 1", direction =-1) +
    coord_fixed(ratio=1) +
    dark_theme_grey(base_size = 12) +
    theme(legend.position = 'none')
}
#------------------------------------------------------------------------------
# RDA Site Ellipse plot function
#------------------------------------------------------------------------------
# With an ordination via redundancy analysis, the positions of species should be interpreted as arrows. It is straightforward to show the positions of the species as arrows. However, with a large number of species, it is better to reduce the number to those that explain more of the variation among the sites. This information is obtained with the envfit function of vegan.

RDA_spe <- function(rda_out) {
  # -----------------------------------------------------------------------
  # Index items from rda_output
  axis_long <- rda_out[[1]] %>% as_tibble()
  sites_long <- rda_out[[2]] %>% as_tibble() %>% fix_site_order() %>% add_rain()
  species_long <- rda_out[[3]] %>% as_tibble()
  RDA_ev <- rda_out[[4]] %>% as_tibble()
  RDA_sv <- rda_out[[5]] %>% as_tibble()
  mycent_alltime <- rda_out[[7]] %>% as_tibble()
  labs_alltime <- rda_out[[8]] %>% as_tibble()
  
  # -----------------------------------------------------------------------
  # Species Plot
  ggplot() +
    # stat_ellipse(data = sites_long,
    #              aes(x=axis1, y=axis2, fill = as.factor(annualrain)),
    #              geom='polygon', alpha=.2, level = 0.9,show.legend = F) +
    # stat_ellipse(data = sites_long,
    #              aes(x=axis1, y=axis2, color = as.factor(annualrain)),
    #              alpha=.7, level = 0.9, show.legend = F) +
    #geom_rect(aes(ymin=axis2_lwr, ymax=axis2_upr,
    #              xmin=axis1_lwr, xmax=axis1_upr,
    #              fill=annualrain),
    #          data= mycent_alltime, 
    #          color=NA, alpha = .25) +
    geom_point(data=species_long,
               aes(x=axis1, y=axis2), 
               color = 'white', fill='black', shape = 24, size = 2) +
    geom_label_repel(data=RDA_sv,
                     aes(x=axis1, y=axis2, label=labels),
                     colour="white", fill = 'grey15', box.padding = .9,
                     max.overlaps = 25) +
    geom_label(aes(label=annualrain_label,
                   x=axis1_mu, y=axis2_mu, fill = as.factor(annualrain)),
               data= labs_alltime, show.legend=T, color ='black' ) +
    xlab(axis_long[1, "label"]) +
    ylab(axis_long[2, "label"]) +
    scale_x_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
    scale_y_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
    scale_fill_manual(values = my_colors) +
    scale_color_manual(values=my_colors) +
    #paletteer::scale_fill_paletteer_c("grDevices::Zissou 1", direction=-1) +
    #paletteer::scale_color_paletteer_c("grDevices::Zissou 1", direction =-1) +
    coord_fixed(ratio=1) +
    dark_theme_grey(base_size = 12) +
    theme(legend.position = 'none')
}

#------------------------------------------------------------------------------
# RDA time trace
#------------------------------------------------------------------------------
RDA_timetrack <- function(rda_out, my_before, my_after) {
  
  # -----------------------------------------------------------------------
  # Index items from rda_output
  axis_long <- rda_out[[1]] %>% as_tibble()
  sites_long <- rda_out[[2]] %>% as_tibble()
  species_long <- rda_out[[3]] %>% as_tibble()
  RDA_ev <- rda_out[[4]] %>% as_tibble()
  RDA_sv <- rda_out[[5]] %>% as_tibble()
  mycent_alltime <- rda_out[[7]] %>% as_tibble()
  labs_alltime <- rda_out[[8]] %>% as_tibble()
  
  # ---------------------------------------------------------------------------
  # plot
  sites_long %>%
    add_rain() %>%
    filter(collection_period < ymd(my_before) & 
             collection_period > ymd(my_after)) %>%
    ggplot() +
    geom_text_repel(data=RDA_sv,
                    aes(x=axis1, y=axis2, label=labels),
                    color="grey85", parse=T) +
    geom_path(aes(x=axis1*.97, y=axis2*.95, color = as.factor(annualrain)),
              arrow = arrow(length = unit(0.2, "inches"),
                            angle = 15, ends = "last", type = "closed"),
              linetype=1) +
    geom_label(aes(x = axis1, y = axis2, label = months_to_hh, 
                   color = as.factor(annualrain)), fill = 'black') +
    geom_label(aes(x = axis1_mu, y = axis2_mu, 
                   label = annualrain_label, fill = as.factor(annualrain)), 
               color = 'black', data = labs_alltime) +
    scale_color_manual(values = my_colors) +
    scale_fill_manual(values = my_colors) +
    dark_theme_grey(base_size = 12) +
    xlab(axis_long[1, "label"]) +
    ylab(axis_long[2, "label"]) +
    theme(legend.position = 'none')
}

#------------------------------------------------------------------------------
# RDA recovery vs baseline
#------------------------------------------------------------------------------
RDA_recovery <- function(rda_out) {
  
  # -----------------------------------------------------------------------
  # Index items from rda_output
  axis_long <- rda_out[[1]] %>% as_tibble()
  sites_long <- rda_out[[2]] %>% as_tibble() %>% add_rain() %>% fix_site_order()
  species_long <- rda_out[[3]] %>% as_tibble()
  RDA_ev <- rda_out[[4]] %>% as_tibble()
  RDA_sv <- rda_out[[5]] %>% as_tibble()
  mycent_alltime <- rda_out[[7]] %>% as_tibble()
  labs_alltime <- rda_out[[8]] %>% as_tibble()
  
  recovery_start = ymd("2017-08-27")
  recovery_end = ymd("2018-05-01")
  
  # ---------------------------------------------------------------------------
  # plot
  d_prep <- sites_long %>%
    mutate(H_effect = ifelse(collection_period > recovery_start & 
                               collection_period < recovery_end, 
                             "Recovery", "Baseline")) %>%
    create_site_group()
  
  d_recovery <- filter(d_prep, H_effect == 'Recovery')
  d_baseline <- filter(d_prep, H_effect == 'Baseline')
  
  d_prep %>%
    ggplot() +
    facet_wrap(~site_group, ncol=1) +
    scale_color_manual(values = my_colors) +
    scale_fill_manual(values = my_colors) +
    scale_shape_manual(values = c(22,21)) +
    dark_theme_grey(base_size = 12) +
    xlab(axis_long[1, "label"]) +
    ylab(axis_long[2, "label"]) +
    geom_point(aes(x = axis1, y = axis2,
                   color = site_code,
                   shape=H_effect), size = 4, show.legend = T) +
    geom_point(data = d_baseline,
               aes(x = axis1, y = axis2, fill=site_code, color=site_code), 
               size = 4, shape=22, show.legend = F)
}
#------------------------------------------------------------------------------
# RDA all_time
#------------------------------------------------------------------------------
rda_all <- my_rda(d_env = d_predictors,
                  d_spec = d_spe)

rda_all_spe_plot <-  rda_all %>% RDA_spe()

rda_all_env_plot <- rda_all %>% RDA_env()

rda_all_hh_plot <- RDA_recovery(rda_all)

rda_all_time_1_plot <- rda_all %>%
  RDA_timetrack(my_before = '2018-03-01', my_after = '2017-01-01')

rda_all_time_2_plot <- rda_all %>%
  RDA_timetrack(my_before = '2018-09-01', my_after = '2018-03-01') 

rda_all_time_3_plot <- rda_all %>%
  RDA_timetrack(my_before = '2020-01-01', my_after = '2018-09-01')

rda_all_time_4_plot <- rda_all %>%
  RDA_timetrack(my_before = '2021-01-01', my_after = '2020-01-01')

#------------------------------------------------------------------------------
# Figures: Biomass-Based (2017-2018)
#------------------------------------------------------------------------------
trim_dates <- function(my_data, after, before) {
  my_data %>%
    rownames_to_column('site_date') %>%
    as_tibble() %>%
    separate(site_date, into=c('site_code', 'collection_period'), sep='_') %>%
    mutate(collection_period = ymd(collection_period)) %>%
    filter(collection_period > ymd(after) &
             collection_period < ymd(before)) %>%
    mutate(UID = paste(site_code, collection_period, sep = '_')) %>%
    select(!c(site_code, collection_period)) %>%
    column_to_rownames('UID')
}

d_predictors_1718 <- trim_dates(d_predictors, '2017-01-01', '2019-01-01')
d_spe_1718 <- trim_dates(d_spe, '2017-01-01', '2019-01-01')

rda_1718 <- my_rda(d_env = d_predictors_1718,
                   d_spec = d_spe_1718)

rda_1718_spe_plot <-  rda_1718 %>% RDA_spe()

rda_1718_env_plot <- rda_1718 %>% RDA_env()

rda_1718_hh_plot <- RDA_recovery(rda_1718)

rda_1718_time_1_plot <- rda_1718 %>%
  RDA_timetrack(my_before = '2018-03-01', my_after = '2017-01-01')

rda_1718_time_2_plot <- rda_1718 %>%
  RDA_timetrack(my_before = '2019-09-01', my_after = '2018-03-01') 

#------------------------------------------------------------------------------
# Figures: (2019-2020)
#------------------------------------------------------------------------------
d_predictors_1920 <- trim_dates(d_predictors, '2019-01-01', '2021-01-01')
d_spe_1920 <- trim_dates(d_spe, '2019-01-01', '2021-01-01')

rda_1920 <- my_rda(d_env = d_predictors_1920,
                   d_spec = d_spe_1920)

rda_1920_spe_plot <-  rda_1920 %>% RDA_spe()

rda_1920_env_plot <- rda_1920 %>% RDA_env()

rda_1920_hh_plot <- RDA_recovery(rda_1920)

rda_1920_time_3_plot <- rda_1920 %>%
  RDA_timetrack(my_before = '2021-01-01', my_after = '2019-01-01')

#------------------------------------------------------------------------------
# Export tables
#------------------------------------------------------------------------------
rda_extract <- function(df, item) {

  extract_axis <- df$axis_long %>% as_tibble() %>%
    mutate(x = label) %>%
    separate(x, into=c('xxx1', 'xxx2'), sep=' ') %>%
    mutate(variation_explained = gsub("[(%)]", "", xxx2) %>% as.numeric()) %>%
    select(-contains('xxx'))

  extract_site <- df$sites_long %>%
    as_tibble() %>%
    select(site_code, collection_period, axis1, axis2) %>%
    mutate(collection_period = as_date(collection_period)) %>%
    left_join(rda_all$centroid_distance %>%
                as_tibble() )

  extract_vector <- df$RDA_table %>%
    rename(p = `p-value`,
           axis1 = `Axis-1`,
           axis2 = `Axis-2`) %>%
    mutate(type='physical') %>%
    full_join(rda_all$RDA_sv %>% rename(Vector = labels) %>%
                mutate(type='biological'))

  if (item == 'axis') { return(extract_axis)}
  else { if (item == 'site') {return(extract_site)}
    else {return(extract_vector)}}
  }

rda_all_extract_axis <- rda_extract(rda_all, item = 'axis')
rda_all_extract_site <- rda_extract(rda_all, item = 'site')
rda_all_extract_vector <- rda_extract(rda_all, item = 'vector')

rda_1718_extract_axis <- rda_extract(rda_1718, item = 'axis')
rda_1718_extract_site <- rda_extract(rda_1718, item = 'site')
rda_1718_extract_vector <- rda_extract(rda_1718, item = 'vector')

rda_1920_extract_axis <- rda_extract(rda_1920, item = 'axis')
rda_1920_extract_site <- rda_extract(rda_1920, item = 'site')
rda_1920_extract_vector <- rda_extract(rda_1920, item = 'vector')

my_objects <- ls()
my_csv_names <- my_objects[str_detect(my_objects, '_extract_')]

my_tables <- list(
  rda_1718_extract_axis, rda_1718_extract_site, rda_1718_extract_vector,
  rda_1920_extract_axis, rda_1920_extract_site, rda_1920_extract_vector,
  rda_all_extract_axis, rda_all_extract_site, rda_all_extract_vector)

names(my_tables) <- my_csv_names

for (i in 1:length(my_tables)) {
  my_place <- paste('exploration/output/', names(my_tables[i]), ".csv", sep='')
  my_object <- my_tables[[i]]
  write_csv(my_object, my_place) }

# #------------------------------------------------------------------------------
# # Export figures
# #------------------------------------------------------------------------------
# my_figure_names <- my_objects[str_detect(my_objects, '_plot')]
# 
# my_figures <- list(
#   rda_1718_env_plot, rda_1718_hh_plot, rda_1718_spe_plot, rda_1718_time_1_plot,
#   rda_1718_time_2_plot, rda_1920_env_plot, rda_1920_hh_plot, rda_1920_spe_plot,
#   rda_1920_time_3_plot, rda_all_env_plot, rda_all_hh_plot, rda_all_spe_plot,
#   rda_all_time_1_plot, rda_all_time_2_plot, rda_all_time_3_plot,
#   rda_all_time_4_plot)
# 
# names(my_figures) <- my_figure_names
# 
# for (i in 1:length(my_figures)) {
#   my_place <- paste('exploration/visualization/', names(my_figures[i]), ".png", sep='')
#   my_object <- my_figures[[i]]
#   ggsave(my_place,
#          plot = my_object,
#          width = 10,
#          height = 10,
#          units = c("in")) }

#------------------------------------------------------------------------------
# End Ordination
